BCFKeyValueData: move various helper code into separate file
