+++ CMakeLists.txt
@@ -203,7 +203,8 @@ add_library(glnexus
             include/discovery.h src/discovery.cc
             include/unifier.h src/unifier.cc
             include/genotyper.h src/genotyper.cc
-            include/genotyper_utils.h
+            src/genotyper_utils.h
+            src/BCFKeyValueData_utils.h
             include/residuals.h src/residuals.cc
             include/KeyValue.h src/KeyValue.cc
             include/BCFSerialize.h src/BCFSerialize.cc
